Batching Service (#6)
The Context was refactored into Context and ContextLifeCycle. The LifeCycle defines the expected event flow of the transaction. The send/recv aka unary service is very simple, it expects a request and returns a response. However, for streaming services, you can imagine many different scenarios. On such scenario is a “batching stream”.

A “batching stream” is a stream where the service reads and collects the elements of the stream until the client signifies it is done writing. That is the signal at which YAIS performs a single batched inference call on the concatenated set of requests that came in over the stream. After the inference calculation is complete, the server writes the results for each request item to the stream. That is, for each request that came in on the stream, the server is expected to return a response.

The batching service is essentially a service that conforms to the unary service. But instead of doing the work, it simply batches and forwards to the inference server that implements the batching stream.

There are several advantages for doing a stream instead of sending a single concatenated unary batch:

You could push items on to the stream immediate as they come in rather than holding them until the batch is full or times out. This allows your pre-processors to get a head start. I will be doing this in the future, but for now I’m keeping it simple.
Streams through a load-balancer are balanced on a per stream basis. So each write on the stream is guaranteed to hit the same endpoint.
And finally, the batching service doesn’t need to know anything about the format of the request/response messages. It simply accepts and forwards them. The result is that the batcher should be able to work with any unary service using any request/response message. You simply need to implement a service capable of handling a stream.